+++ readWrite/src/main/scala-2.13+/slinky/readwrite/CompatUtil.scala
@@ -1,0 +1,5 @@
+package slinky.readwrite
+
+object CompatUtil {
+  type Factory[-A, +C] = scala.collection.Factory[A, C]
+}